Add LoadMany and update errors
In order to support LoadMany and make it possible to identify errors by
the original key, revamp the way errors are structured to permit the key
to be recovered.

Uses this to add LoadMany(...K) ([]V, []error) which can return a mix of
results and errors, not necessarily in the order requested.